It would be useful if create_distributed_function would allow you us force pushdown of stored procedure calls even if the procedure is called from a bigger stored procedure/transaction, with the restriction that the inner procedure can only do local execution and will error out if it opens a connection.
(1) Functions will be be delegated even when present in the scope an explicit BEGIN/COMMIT transaction block or UDF calling another UDF.
(2) Prohibit/Limit the delegated function not to do a 2PC (or any work on a remote connection)
(3) Have a safety net around ensuring the (2) i.e. we should block the connections from the delegated procedure or make sure that no 2PC happens on the node.
(4) Such delegated functions should only use the distributed argument value

To limit the scope of the project we are considering only Functions(not procedures) for the initial work.

DESCRIPTION: Introduce a new flag for "forcePushdown" in create_distributed_function(), which will allow the function to be delegated in an explicit transaction block.

Fixes #3265

Once the function is delegated to the worker, on that node during the planning

distributed_planner()
   TryToDelegateFunctionCall()
      CheckDelegatedFunctionExecution()
         EnableInForceDelegatedFuncExecution() Save the distribution argument (Constant)
ExecutorStart()
  CitusBeginScan()
    GetInForceDelegatedFuncExecution()  Make sure to not use non-distribution argument"

ExecutorRun()
  AdaptiveExecutor()
    StartDistributedExecution()
      EnsureDelegatedFunctionSafeExecution()  Make sure all the shards are local to the node in remoteTaskList